cmd/snap: skip tests depending on text wrapping
Upstream snapd conains tests that check the output of various commands
along with the --help command-line argument. The output is wrapped to
match terminal width and for readability. The algorithm for wrapping
has apparently changed across versions of github.com/jessevdk/go-flags.

Since this test is not critical for anything it can be disabled to let
the package build.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>